+++ src/rules/prBodyDescriptive.ts
@@ -1,7 +1,7 @@
-import { wrapSafe } from "../types/utils.js";
+import { findPrTemplate } from "../action/findPrTemplate.js";
 import { defineRule } from "./defineRule.js";
 
-export const prBodyNotEmpty = defineRule({
+export const prBodyDescriptive = defineRule({
 	about: {
 		config: "recommended",
 		description: "PRs should have a description beyond the template.",
@@ -9,7 +9,7 @@ export const prBodyNotEmpty = defineRule({
 			`This repository expects pull requests to include a description explaining the changes.`,
 			`The description should have at least one word not in the PR template, or any content if no template exists.`,
 		],
-		name: "pr-body-not-empty",
+		name: "pr-body-descriptive",
 	},
 	async pullRequest(context, entity) {
 		if (!entity.data.body) {
@@ -22,19 +22,9 @@ export const prBodyNotEmpty = defineRule({
 			return;
 		}
 
-		const templateResponse = await wrapSafe(
-			context.octokit.rest.repos.getContent({
-				owner: context.locator.owner,
-				path: ".github/PULL_REQUEST_TEMPLATE.md",
-				repo: context.locator.repository,
-			}),
-		);
+		const template = await findPrTemplate(context.octokit, context.locator);
 
-		if (
-			!templateResponse ||
-			Array.isArray(templateResponse.data) ||
-			templateResponse.data.type !== "file"
-		) {
+		if (!template) {
 			if (
 				entity.data.body
 					.trim()
@@ -50,11 +40,6 @@ export const prBodyNotEmpty = defineRule({
 			}
 			return;
 		}
-
-		const template = Buffer.from(
-			templateResponse.data.content,
-			"base64",
-		).toString("utf-8");
 
 		const templateWords = new Set(
 			template.toLowerCase().match(/[\p{L}\p{N}]+/gu) ?? [],