+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+# [0.14.0](https://github.com/JoshuaKGoldberg/OctoGuide/compare/0.13.0...0.14.0) (2025-08-22)
+
+### Features
+
+- add config none ([#242](https://github.com/JoshuaKGoldberg/OctoGuide/issues/242)) ([1863dab](https://github.com/JoshuaKGoldberg/OctoGuide/commit/1863dab41c3fd8d12c84414216e0ae28778646f7)), closes [#241](https://github.com/JoshuaKGoldberg/OctoGuide/issues/241)
+
+# [0.13.0](https://github.com/JoshuaKGoldberg/OctoGuide/compare/0.12.1...0.13.0) (2025-08-22)
+
+### Features
+
+- mark passed comments as resolved ([#238](https://github.com/JoshuaKGoldberg/OctoGuide/issues/238)) ([58ec5c3](https://github.com/JoshuaKGoldberg/OctoGuide/commit/58ec5c383d8c63187991ae236e827f0c5bb8ae2a)), closes [#202](https://github.com/JoshuaKGoldberg/OctoGuide/issues/202)
+
+## [0.12.1](https://github.com/JoshuaKGoldberg/OctoGuide/compare/0.12.0...0.12.1) (2025-08-19)
+
+### Bug Fixes
+
+- support all github pr template locations ([#99](https://github.com/JoshuaKGoldberg/OctoGuide/issues/99)) ([04ca8bb](https://github.com/JoshuaKGoldberg/OctoGuide/commit/04ca8bb44b2730032f21038ac3d9dce20c79756a)), closes [#47](https://github.com/JoshuaKGoldberg/OctoGuide/issues/47)
+
 # [0.12.0](https://github.com/JoshuaKGoldberg/OctoGuide/compare/0.11.4...0.12.0) (2025-07-24)
 
 ### Features